+++ ui/src/pages/chat/chat-history.ts
@@ -15,7 +15,11 @@ import {
   isAssistantHeartbeatAckForDisplay,
   stripHeartbeatTokenForDisplay,
 } from "../../lib/chat/heartbeat-display.ts";
-import { extractText, hasTranscriptMediaFacts } from "../../lib/chat/message-extract.ts";
+import {
+  extractText,
+  hasTranscriptMediaFacts,
+  isEmptyUserTextOnlyMessage,
+} from "../../lib/chat/message-extract.ts";
 import {
   retirePendingChatSideQuestion,
   type ChatSideResult,
@@ -179,50 +183,6 @@ function isSyntheticTranscriptRepairToolResult(message: unknown): boolean {
   }
   const text = extractText(message);
   return typeof text === "string" && text.trim() === SYNTHETIC_TRANSCRIPT_REPAIR_RESULT;
-}
-
-function isTextOnlyContent(content: unknown): boolean {
-  if (typeof content === "string") {
-    return true;
-  }
-  if (!Array.isArray(content)) {
-    return false;
-  }
-  if (content.length === 0) {
-    return true;
-  }
-  let sawText = false;
-  for (const block of content) {
-    if (!block || typeof block !== "object") {
-      return false;
-    }
-    const entry = block as { type?: unknown; text?: unknown };
-    if (entry.type !== "text") {
-      return false;
-    }
-    sawText = true;
-    if (typeof entry.text !== "string") {
-      return false;
-    }
-  }
-  return sawText;
-}
-
-export function isEmptyUserTextOnlyMessage(message: unknown): boolean {
-  if (!message || typeof message !== "object") {
-    return false;
-  }
-  const entry = message as Record<string, unknown>;
-  if (normalizeLowercaseStringOrEmpty(entry.role) !== "user") {
-    return false;
-  }
-  if (hasTranscriptMediaFacts(entry)) {
-    return false;
-  }
-  if (!isTextOnlyContent(entry.content ?? entry.text)) {
-    return false;
-  }
-  return (extractText(message)?.trim() ?? "") === "";
 }
 
 function isHeartbeatAckStream(text: string): boolean {
